add error checking for 64 bit textures

+++ GW Launcher/uMod/TexBundle.cs
@@ -23,7 +23,7 @@ public class TexBundle
         foreach (var (crc, value) in loader.Entries)
         {
             TexDef def;
-            def.crcHash = (uint) Convert.ToUInt64(crc, 16);
+            def.crcHash = Convert.ToUInt32(crc, 16);
             def.fileData = value;
 
             defs.Add(def);
